requirements.txt
python-telegram-bot[webhooks]==21.3
httpx==0.27.2
asyncpg==0.29.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
uvloop==0.19.0
qrcode[pil]==7.4.2
aiohttp==3.9.5
